Move NativeServicesTestFixture to internal-testing
